Move common/histogram to internal/.
