+++ cartographer/mapping/internal/3d/pose_graph/constraint_builder_3d.h
@@ -26,7 +26,7 @@
 #include "Eigen/Core"
 #include "Eigen/Geometry"
 #include "cartographer/common/fixed_ratio_sampler.h"
-#include "cartographer/common/histogram.h"
+#include "cartographer/common/internal/histogram.h"
 #include "cartographer/common/lua_parameter_dictionary.h"
 #include "cartographer/common/math.h"
 #include "cartographer/common/mutex.h"
